Rename ad -> nad
