+++ network-area-diagram-viewer/src/main/java/com/powsybl/nad/viewer/NadViewerLauncher.java
@@ -4,9 +4,8 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
-package com.powsybl.ad.viewer.main;
+package com.powsybl.nad.viewer;
 
-import com.powsybl.ad.viewer.view.AreaDiagramViewer;
 import javafx.application.Application;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -14,15 +13,15 @@ import org.slf4j.LoggerFactory;
 /**
  * @author Louis Lhotte <louis.lhotte@student-cs.fr>
  */
-public class Main {
+public class NadViewerLauncher {
 
-    private Main() {
+    private NadViewerLauncher() {
     }
 
-    private static final Logger LOGGER = LoggerFactory.getLogger(Main.class);
+    private static final Logger LOGGER = LoggerFactory.getLogger(NadViewerLauncher.class);
 
     public static void main(String[] args) {
         LOGGER.info("Starting Area Diagram Viewer application.");
-        Application.launch(AreaDiagramViewer.class);
+        Application.launch(NetworkAreaDiagramViewer.class);
     }
 }
