+++ network-area-diagram-viewer/src/main/java/com/powsybl/nad/viewer/controller/ControllerOptions.java
@@ -4,13 +4,12 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
-package com.powsybl.ad.viewer.controller;
+package com.powsybl.nad.viewer.controller;
 
-import com.powsybl.ad.viewer.model.NadCalls;
-import com.powsybl.ad.viewer.util.Util;
-import com.powsybl.ad.viewer.view.OptionsPane;
-import com.powsybl.ad.viewer.view.diagram.DiagramPane;
-import com.powsybl.ad.viewer.view.diagram.containers.*;
+import com.powsybl.nad.viewer.model.NadCalls;
+import com.powsybl.nad.viewer.view.OptionsPane;
+import com.powsybl.nad.viewer.view.diagram.DiagramPane;
+import com.powsybl.nad.viewer.view.diagram.containers.*;
 import com.powsybl.iidm.network.Container;
 import com.powsybl.iidm.network.Network;
 import com.powsybl.iidm.network.Substation;
@@ -32,8 +31,8 @@ import java.util.Map;
 import java.util.function.Function;
 import java.util.stream.Collectors;
 
-import static com.powsybl.ad.viewer.controller.ControllerParameters.getParamPane;
-import static com.powsybl.ad.viewer.model.NadCalls.*;
+import static com.powsybl.nad.viewer.controller.ControllerParameters.getParamPane;
+import static com.powsybl.nad.viewer.model.NadCalls.*;
 
 /**
  * @author Louis Lhotte <louis.lhotte@student-cs.fr>
