+++ network-area-diagram-viewer/src/main/java/com/powsybl/nad/viewer/view/diagram/containers/ContainerDiagramPane.java
@@ -4,7 +4,7 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
-package com.powsybl.ad.viewer.view.diagram.containers;
+package com.powsybl.nad.viewer.view.diagram.containers;
 
 import javafx.geometry.Side;
 import javafx.scene.control.Tab;
