+++ network-area-diagram-viewer/src/main/java/com/powsybl/nad/viewer/view/diagram/containers/SelectableSubstation.java
@@ -1,4 +1,4 @@
-package com.powsybl.ad.viewer.view.diagram.containers;
+package com.powsybl.nad.viewer.view.diagram.containers;
 
 public class SelectableSubstation extends AbstractSelectableContainer {
     public SelectableSubstation(String id, String name) {
