+++ network-area-diagram-viewer/src/main/resources/images/logo.png
